update applications
Former-commit-id: 23b1510124076b40ed919f40a0ee3a30a1db24b7 [formerly c89b6695499fc33187c4d6aac76ca7a9d8f47418]
Former-commit-id: ea5252ab416ab96cef5858235f1013afb7e528c7

+++ src/Acquisition.hpp
@@ -15,7 +15,7 @@ class SRC_API Measure
 {
 
   public:
-    Measure( const unsigned char* const data, size_t size, bool floatData = false );
+    Measure( const unsigned char* const data, uint64_t size, bool floatData = false );
     Measure( Measure&& measure );
     //    Measure( Measurement measurement );
     Measure( const Measure& ) = delete;
